Duplicate testing function
Unfortunately testing for duplicate links slow down my program a lot
because each new link have to be save by his own and it isn't possible
to save whole link packages anymore but I think that this function is so
important that it must be in the program. Maybe I’ll find a way to save
whole link packages anyway.